Move get_explicit_model_from_cs() Method to utils folder, Include smack foulder output as parameter
